requirements.txt
xmltodict==0.12.0
Flask-Migrate==2.5.3
Flask==1.1.2
pymysql==0.9.3
mysqlclient==1.4.6
marshmallow-sqlalchemy==0.23.1
flask-sqlalchemy==2.4.3
flask-sqlacodegen==1.1.7